requirements.txt
streamlit
openai
firebase-admin
python-dotenv
pandas
requests